refactor(cypher): extract bounded reentry glue from lowering.py (#1295)
Moves 16 reentry helpers (~470 LOC) out of cypher/lowering.py into a focused
cypher/reentry/ subpackage:
- naming.py — hidden-column / carry-name conventions (4 helpers)
- scope.py  — alias-scope traversal for hidden reentry refs (2 helpers)
- carry.py  — prefix carry-column / order helpers (4 helpers)
- rewrite.py — AST/query rewriters for reentry expressions (7 helpers)

All names are re-exported from cypher/lowering.py so existing callers keep
working unchanged. The moved modules import lowering lazily inside function
bodies for the few helpers (_unsupported, _render_expr_node,
_rewrite_where_clause_and_resync, _first_pattern_node_alias, etc.) they
still need from lowering, avoiding circular imports at module load.

lowering.py: 10192 → 9684 LOC (-508).

The orchestrator (_compile_bounded_reentry_query) and
_map_terminal_reentry_query intentionally stay in lowering.py for this slice
— they share heavy edit surface with the sibling #1294 dual-contract
retirement (PR #1297). A follow-on slice can move them once #1297 lands.

Refs: #1260 S2
