+++ CHANGELOG.md
@@ -39,6 +39,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Internal
 - **GFQL / Cypher reentry cleanup — retire dual-contract scalar reentry extras path (#1294, #1260 S1)**: Removed legacy `CompiledCypherExecutionExtras` fields `scalar_reentry_alias` / `scalar_reentry_columns` and migrated runtime/compile branching to the canonical `reentry_plan` contract (`scalar_only` / `free_form` / whole-row lanes). `_execute_compiled_query_with_reentry` now dispatches scalar-only reentry via `ReentryPlan.scalar_columns` instead of legacy fields, lowering extras threading no longer forwards legacy scalars through `_execution_extras_with`, and whole-row reentry dispatch now requires a `ReentryPlan` (retiring the residual projection-contract fallback path). Added regression coverage locking absence of legacy extras fields and scalar-column plan parity in `test_lowering.py`.
 - **GFQL type-system T1 — schema contract + canonical naming alignment (#1296)**: Stabilized the IR schema-catalog contract in `graphistry/compute/gfql/ir/compilation.py` while preserving existing `GFQLSchema is GraphSchemaCatalog` alias behavior. Added `GraphSchemaCatalog.from_schema_parts(...)` to normalize iterable inputs into frozen-set contract shapes and copy metadata safely, plus canonical accessors (`node_id`, `edge_source`, `edge_destination`) and membership helpers (`has_node_column`, `has_edge_column`) to provide a minimal stable interface for follow-on T2/T3 schema-validation and type-propagation lanes under #1262/#1259. Added focused contract tests in `graphistry/tests/compute/gfql/test_ir_compilation.py` and validated adjacent M0B conformance coverage remains green.
+- **GFQL / Cypher monolith shrinkdown — extract bounded reentry glue from `lowering.py` (#1295, #1260 S2)**: Moved 17 reentry helpers (16 public + carry-internal `_literal_limit_value`, ~470 LOC) out of `graphistry/compute/gfql/cypher/lowering.py` into a new `graphistry/compute/gfql/cypher/reentry/` subpackage with four focused modules: `naming.py` (hidden-column / carry-name conventions), `scope.py` (alias-scope traversal for hidden reentry references), `carry.py` (prefix carry-column / order helpers, including `_literal_limit_value`), and `rewrite.py` (AST/query rewriters that retarget reentry expressions onto carried columns). The 16 public names are re-exported from `lowering.py` so existing callers (`from graphistry.compute.gfql.cypher.lowering import _reentry_hidden_column_name`) keep working unchanged; the moved modules pull lowering-internal helpers (`_unsupported`, `_unsupported_at_span`, `_render_expr_node`, `_rebuild_expr_node`, `_rewrite_expr_identifiers`, `_rewrite_where_clause_and_resync`, `_first_pattern_node_alias`, `_parse_row_expr`) lazily inside function bodies to avoid circular imports at module load. `carry.py` declares `ResultProjectionPlan` as a `TYPE_CHECKING`-only forward-string back-edge to lowering. `lowering.py` shrinks materially after the move; the orchestrator (`_compile_bounded_reentry_query`) and `_map_terminal_reentry_query` stay in `lowering.py` and are deferred to a follow-on monolith-shrinkdown slice.
 - **GFQL / Cypher lowering — admit free-form carried-property bridge WHERE+RETURN shape (#1275)**: Extended bounded reentry lowering to admit intermediate free-form `MATCH` shapes where trailing predicates and projections reference carried alias properties (for example `WHERE c.id = x.id` and `RETURN x.id`) by wiring `_rewrite_multi_whole_row_prefix(...)` into the free-form lane, threading `multi_alias_carries` into `ReentryPlan.aliases` + expression rewrite, and removing the prior conservative carried-property failfast. Also fixed a structured-WHERE rewrite gap in reentry lowering: reentry `WhereClause` instances with `expr_tree=None` (predicate-only form) are now rewritten/resynced instead of being skipped/dropped, preventing runtime alias-validation failures on carried-property bridge predicates. Added/converted positive execution tests (including cuDF parity guard) for bridge WHERE and bridge WHERE+RETURN variants in `test_lowering.py` while preserving adjacent #1263 regression coverage.
 - **GFQL / Cypher runtime — admit multi-prefix-row free-form intermediate MATCH (#1285, #1263 follow-up)**: Replaces the single-row-only failfast in `_compiled_query_freeform_reentry_state` with a per-row union loop in `_execute_compiled_query_with_reentry`, mirroring the scalar-only multi-row pattern from #1047. New `_freeform_broadcast_row_to_nodes` helper isolates per-row hidden-carry broadcast; the outer dispatcher loops over prefix rows, runs the suffix as a global MATCH per row, and unions results via the existing engine-polymorphic `_union_scalar_reentry_results`. Cartesian semantics fall out naturally — every prefix row × every trailing-MATCH row appears in the output. Out of scope (preserved as failfasts): `optional_reentry + multi-row free-form` (mirrors the scalar lane's existing guard at #1047). Carried-property bridge admit landed separately in #1287. Tests retargeted: the prior `..._failfast_rejects_..._on_multi_row_prefix` is now `..._executes_..._on_multi_row_prefix` (positive). New tests cover Cartesian semantics (2 prefix × 2 trailing = 4 rows; each pair appears twice), cuDF parity, and the OPTIONAL MATCH failfast. Pre-existing simple (single-row) free-form admits from #1279 stay green.
 - **GFQL / Cypher lowering — #1273 acceptance slice admits scalar multi-alias WITH shape A**: First-stage bounded reentry now admits `MATCH ... WITH a.id AS a_id, b.id AS b_id ... MATCH ...` multi-alias scalar/property projections (shape A) while keeping whole-row multi-alias carries (for example `WITH a, b`) outside the admitted lane. The admission gate is explicitly narrowed to scalar/property projections only, so known TCK xfail contract shape `WITH n, x` (`with-where3-3`) remains fail-fast and unchanged pending broader #1273 closure. Tests were converted to positive execution assertions for the admitted shape, and remaining rejected shapes continue to point explicitly to #1273.
